tidy up
License: MIT
