+++ src/lib/cbor.js
@@ -28,3 +28,16 @@ export function findFirstLinkInPath (node, path) {
   }
   return link
 }
+
+export async function findCid (getIpfs, node, rootCid, rest) {
+  // until ipfs.dag.resolve is available, we have to walk the path to find the nearest cid.
+  // dag.resolve https://github.com/ipfs/js-ipfs-api/pull/755#issuecomment-386882099
+  const firstLinkCid = findFirstLinkInPath(node, rest)
+  if (!firstLinkCid) {
+    // we're in the right node.
+    return rootCid
+  } else {
+    const {value: nextNode} = await getIpfs().dag.get(firstLinkCid)
+    return findCid(getIpfs, nextNode, firstLinkCid, rest)
+  }
+}
